Changes to PR to complete review
- Move CustomHighlight to its own component
- Move safeEscapeString into its own utility function
- Don't run the highlight on the package name
